requirements.txt
python-dotenv==1.0.1
openai==1.47.1
Pillow==10.4.0